Change completude checking
The fact that a line is terminated (by a line delimiter or the end of
the file) has been put in ft_read_file, in order to avoid costly
reallocation of memory. To avoid rescan all the string (which has
already been, only the added part (which is the buffer where read puts
what has been read from the file) is searched.